+++ src/roman_datamodels/maker_utils.py
@@ -458,7 +458,7 @@ def add_ref_common(meta):
     meta["reftype"] = ""
 
 
-def mk_level1_science_raw(shape=None, filepath=None):
+def mk_level1_science_raw(shape=(8, 4096, 4096), filepath=None):
     """
     Create a dummy level 1 ScienceRaw instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -482,16 +482,12 @@ def mk_level1_science_raw(shape=None, filepath=None):
     wfi_science_raw = stnode.WfiScienceRaw()
     wfi_science_raw["meta"] = meta
 
-    if not shape:
-        shape = (8, 4096, 4096)
-        n_ints = 8
-    else:
-        n_ints = shape[0]
+    n_groups = shape[0]
 
     wfi_science_raw["data"] = u.Quantity(np.zeros(shape, dtype=np.uint16), u.DN, dtype=np.uint16)
 
     # add amp 33 ref pix
-    wfi_science_raw["amp33"] = u.Quantity(np.zeros((n_ints, 4096, 128), dtype=np.uint16), u.DN, dtype=np.uint16)
+    wfi_science_raw["amp33"] = u.Quantity(np.zeros((n_groups, 4096, 128), dtype=np.uint16), u.DN, dtype=np.uint16)
 
     if filepath:
         af = asdf.AsdfFile()
@@ -501,7 +497,7 @@ def mk_level1_science_raw(shape=None, filepath=None):
         return wfi_science_raw
 
 
-def mk_level2_image(shape=None, n_ints=None, filepath=None):
+def mk_level2_image(shape=(4088, 4088), n_groups=8, filepath=None):
     """
     Create a dummy level 2 Image instance (or file) with arrays and valid values
     for attributes required by the schema.
@@ -517,9 +513,9 @@ def mk_level2_image(shape=None, n_ints=None, filepath=None):
         the border reference pixel arrays will have (y, x) dimensions (14, 4)
         and (4, 14)). Default is 4088 x 4088.
 
-    n_ints : int
+    n_groups : int
         The level 2 file is flattened, but it contains arrays for the original
-        reference pixels which remain 3D. n_ints specifies what the z dimension
+        reference pixels which remain 3D. n_groups specifies what the z dimension
         of these arrays should be. Defaults to 8.
 
     filepath : str
@@ -533,16 +529,16 @@ def mk_level2_image(shape=None, n_ints=None, filepath=None):
     meta["photometry"] = mk_photometry()
     wfi_image = stnode.WfiImage()
     wfi_image["meta"] = meta
-    if not shape:
-        shape = (4088, 4088)
-    if not n_ints:
-        n_ints = 8
 
     # add border reference pixel arrays
-    wfi_image["border_ref_pix_left"] = u.Quantity(np.zeros((n_ints, shape[0] + 8, 4), dtype=np.float32), u.DN, dtype=np.float32)
-    wfi_image["border_ref_pix_right"] = u.Quantity(np.zeros((n_ints, shape[0] + 8, 4), dtype=np.float32), u.DN, dtype=np.float32)
-    wfi_image["border_ref_pix_top"] = u.Quantity(np.zeros((n_ints, shape[0] + 8, 4), dtype=np.float32), u.DN, dtype=np.float32)
-    wfi_image["border_ref_pix_bottom"] = u.Quantity(np.zeros((n_ints, shape[0] + 8, 4), dtype=np.float32), u.DN, dtype=np.float32)
+    wfi_image["border_ref_pix_left"] = u.Quantity(np.zeros((n_groups, shape[0] + 8, 4), dtype=np.float32), u.DN, dtype=np.float32)
+    wfi_image["border_ref_pix_right"] = u.Quantity(
+        np.zeros((n_groups, shape[0] + 8, 4), dtype=np.float32), u.DN, dtype=np.float32
+    )
+    wfi_image["border_ref_pix_top"] = u.Quantity(np.zeros((n_groups, shape[0] + 8, 4), dtype=np.float32), u.DN, dtype=np.float32)
+    wfi_image["border_ref_pix_bottom"] = u.Quantity(
+        np.zeros((n_groups, shape[0] + 8, 4), dtype=np.float32), u.DN, dtype=np.float32
+    )
 
     # and their dq arrays
     wfi_image["dq_border_ref_pix_left"] = np.zeros((shape[0] + 8, 4), dtype=np.uint32)
@@ -551,7 +547,7 @@ def mk_level2_image(shape=None, n_ints=None, filepath=None):
     wfi_image["dq_border_ref_pix_bottom"] = np.zeros((4, shape[1] + 8), dtype=np.uint32)
 
     # add amp 33 ref pixel array
-    amp33_size = (n_ints, 4096, 128)
+    amp33_size = (n_groups, 4096, 128)
     wfi_image["amp33"] = u.Quantity(np.zeros(amp33_size, dtype=np.uint16), u.DN, dtype=np.uint16)
 
     wfi_image["data"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.electron / u.s, dtype=np.float32)
@@ -618,7 +614,7 @@ def mk_level3_mosaic(shape=None, n_images=2, filepath=None):
         return wfi_mosaic
 
 
-def mk_flat(shape=None, filepath=None):
+def mk_flat(shape=(4096, 4096), filepath=None):
     """
     Create a dummy Flat instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -641,9 +637,6 @@ def mk_flat(shape=None, filepath=None):
     meta["reftype"] = "FLAT"
     flatref["meta"] = meta
 
-    if not shape:
-        shape = (4096, 4096)
-
     flatref["data"] = np.zeros(shape, dtype=np.float32)
     flatref["dq"] = np.zeros(shape, dtype=np.uint32)
     flatref["err"] = np.zeros(shape, dtype=np.float32)
@@ -656,7 +649,7 @@ def mk_flat(shape=None, filepath=None):
         return flatref
 
 
-def mk_dark(shape=None, filepath=None):
+def mk_dark(shape=(2, 4096, 4096), filepath=None):
     """
     Create a dummy Dark Current instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -687,9 +680,6 @@ def mk_dark(shape=None, filepath=None):
     exposure["ma_table_name"] = NOSTR
     exposure["ma_table_number"] = NONUM
     darkref["meta"]["exposure"] = exposure
-
-    if not shape:
-        shape = (2, 4096, 4096)
 
     darkref["data"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.DN, dtype=np.float32)
     darkref["dq"] = np.zeros(shape[1:], dtype=np.uint32)
@@ -737,7 +727,7 @@ def mk_distortion(filepath=None):
         return distortionref
 
 
-def mk_gain(shape=None, filepath=None):
+def mk_gain(shape=(4096, 4096), filepath=None):
     """
     Create a dummy Gain instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -760,9 +750,6 @@ def mk_gain(shape=None, filepath=None):
     meta["reftype"] = "GAIN"
     gainref["meta"] = meta
 
-    if not shape:
-        shape = (4096, 4096)
-
     gainref["data"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.electron / u.DN, dtype=np.float32)
 
     if filepath:
@@ -773,7 +760,7 @@ def mk_gain(shape=None, filepath=None):
         return gainref
 
 
-def mk_ipc(shape=None, filepath=None):
+def mk_ipc(shape=(3, 3), filepath=None):
     """
     Create a dummy IPC instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -796,9 +783,6 @@ def mk_ipc(shape=None, filepath=None):
     meta["reftype"] = "IPC"
     ipcref["meta"] = meta
 
-    if not shape:
-        shape = (3, 3)
-
     ipcref["data"] = np.zeros(shape, dtype=np.float32)
     ipcref["data"][int(np.floor(shape[0] / 2))][int(np.floor(shape[1] / 2))] = 1.0
 
@@ -810,7 +794,7 @@ def mk_ipc(shape=None, filepath=None):
         return ipcref
 
 
-def mk_linearity(shape=None, filepath=None):
+def mk_linearity(shape=(2, 4096, 4096), filepath=None):
     """
     Create a dummy Linearity instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -836,9 +820,6 @@ def mk_linearity(shape=None, filepath=None):
     linearityref["meta"]["input_units"] = u.DN
     linearityref["meta"]["output_units"] = u.DN
 
-    if not shape:
-        shape = (2, 4096, 4096)
-
     linearityref["dq"] = np.zeros(shape[1:], dtype=np.uint32)
     linearityref["coeffs"] = np.zeros(shape, dtype=np.float32)
 
@@ -850,7 +831,7 @@ def mk_linearity(shape=None, filepath=None):
         return linearityref
 
 
-def mk_inverse_linearity(shape=None, filepath=None):
+def mk_inverse_linearity(shape=(2, 4096, 4096), filepath=None):
     """
     Create a dummy InverseLinearity instance (or file) with arrays and valid
     values for attributes required by the schema.
@@ -876,9 +857,6 @@ def mk_inverse_linearity(shape=None, filepath=None):
     inverselinearityref["meta"]["input_units"] = u.DN
     inverselinearityref["meta"]["output_units"] = u.DN
 
-    if not shape:
-        shape = (2, 4096, 4096)
-
     inverselinearityref["dq"] = np.zeros(shape[1:], dtype=np.uint32)
     inverselinearityref["coeffs"] = np.zeros(shape, dtype=np.float32)
 
@@ -890,7 +868,7 @@ def mk_inverse_linearity(shape=None, filepath=None):
         return inverselinearityref
 
 
-def mk_mask(shape=None, filepath=None):
+def mk_mask(shape=(4096, 4096), filepath=None):
     """
     Create a dummy Mask instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -913,9 +891,6 @@ def mk_mask(shape=None, filepath=None):
     meta["reftype"] = "MASK"
     maskref["meta"] = meta
 
-    if not shape:
-        shape = (4096, 4096)
-
     maskref["dq"] = np.zeros(shape, dtype=np.uint32)
 
     if filepath:
@@ -926,7 +901,7 @@ def mk_mask(shape=None, filepath=None):
         return maskref
 
 
-def mk_pixelarea(shape=None, filepath=None):
+def mk_pixelarea(shape=(4096, 4096), filepath=None):
     """
     Create a dummy Pixelarea instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -952,9 +927,6 @@ def mk_pixelarea(shape=None, filepath=None):
         "pixelarea_arcsecsq": float(NONUM) * u.arcsec**2,
     }
     pixelarearef["meta"] = meta
-
-    if not shape:
-        shape = (4096, 4096)
 
     pixelarearef["data"] = np.zeros(shape, dtype=np.float32)
 
@@ -1041,7 +1013,7 @@ def mk_wfi_img_photom(filepath=None):
         return wfi_img_photomref
 
 
-def mk_readnoise(shape=None, filepath=None):
+def mk_readnoise(shape=(4096, 4096), filepath=None):
     """
     Create a dummy Readnoise instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -1068,9 +1040,6 @@ def mk_readnoise(shape=None, filepath=None):
     exposure["p_exptype"] = "WFI_IMAGE|WFI_GRISM|WFI_PRISM|"
     readnoiseref["meta"]["exposure"] = exposure
 
-    if not shape:
-        shape = (4096, 4096)
-
     readnoiseref["data"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.DN, dtype=np.float32)
 
     if filepath:
@@ -1081,7 +1050,7 @@ def mk_readnoise(shape=None, filepath=None):
         return readnoiseref
 
 
-def mk_ramp(shape=None, n_ints=None, filepath=None):
+def mk_ramp(shape=(8, 4096, 4096), filepath=None):
     """
     Create a dummy Ramp instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -1089,11 +1058,9 @@ def mk_ramp(shape=None, n_ints=None, filepath=None):
     Parameters
     ----------
     shape : tuple, int
-        (optional) Shape (y, x) of data array in the model (and its
+        (optional) Shape (z, y, x) of data array in the model (and its
         corresponding dq/err arrays). This specified size includes the
-        four-pixel border of reference pixels. Default is 4096 x 4096.
-
-    n_ints : int
+        four-pixel border of reference pixels. Default is 8 x 4096 x 4096.
 
     filepath : str
         (optional) File name and path to write model to.
@@ -1105,9 +1072,6 @@ def mk_ramp(shape=None, n_ints=None, filepath=None):
     meta = mk_common_meta()
     ramp = stnode.Ramp()
     ramp["meta"] = meta
-
-    if not shape:
-        shape = (8, 4096, 4096)
 
     # add border reference pixel arrays
     ramp["border_ref_pix_left"] = u.Quantity(np.zeros((shape[0], shape[1], 4), dtype=np.float32), u.DN, dtype=np.float32)
@@ -1122,7 +1086,7 @@ def mk_ramp(shape=None, n_ints=None, filepath=None):
     ramp["dq_border_ref_pix_bottom"] = np.zeros((4, shape[2]), dtype=np.uint32)
 
     # add amp 33 ref pixel array
-    ramp["amp33"] = u.Quantity(np.full(shape, 1.0, dtype=np.uint16), u.DN, dtype=np.uint16)
+    ramp["amp33"] = u.Quantity(np.zeros((shape[0], shape[1], 128), dtype=np.uint16), u.DN, dtype=np.uint16)
 
     ramp["data"] = u.Quantity(np.full(shape, 1.0, dtype=np.float32), u.DN, dtype=np.float32)
     ramp["pixeldq"] = np.zeros(shape[1:], dtype=np.uint32)
@@ -1137,7 +1101,7 @@ def mk_ramp(shape=None, n_ints=None, filepath=None):
         return ramp
 
 
-def mk_rampfitoutput(shape=None, filepath=None):
+def mk_rampfitoutput(shape=(8, 4096, 4096), filepath=None):
     """
     Create a dummy Rampfit Output instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -1158,9 +1122,6 @@ def mk_rampfitoutput(shape=None, filepath=None):
     rampfitoutput = stnode.RampFitOutput()
     rampfitoutput["meta"] = meta
 
-    if not shape:
-        shape = (8, 4096, 4096)
-
     rampfitoutput["slope"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.electron / u.s, dtype=np.float32)
     rampfitoutput["sigslope"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.electron / u.s, dtype=np.float32)
     rampfitoutput["yint"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.electron, dtype=np.float32)
@@ -1179,7 +1140,7 @@ def mk_rampfitoutput(shape=None, filepath=None):
         return rampfitoutput
 
 
-def mk_associations(shape=None, filepath=None):
+def mk_associations(shape=(2, 3, 1), filepath=None):
     """
     Create a dummy Association table instance (or file) with table and valid values for attributes
     required by the schema.
@@ -1213,9 +1174,6 @@ def mk_associations(shape=None, filepath=None):
     associations["asn_pool"] = "r00001_20200530t023154_pool"
     associations["target"] = 16
 
-    if not shape:
-        shape = (2, 3, 1)
-
     file_idx = 0
     associations["products"] = []
     for product_idx in range(len(shape)):
@@ -1236,7 +1194,7 @@ def mk_associations(shape=None, filepath=None):
         return associations
 
 
-def mk_guidewindow(shape=None, filepath=None):
+def mk_guidewindow(shape=(2, 8, 16, 32, 32), filepath=None):
     """
     Create a dummy Guidewindow instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -1280,9 +1238,6 @@ def mk_guidewindow(shape=None, filepath=None):
     guidewindow["meta"]["data_end"] = NONUM
     guidewindow["meta"]["gw_acq_exec_stat"] = generate_string("Status ", 15)
 
-    if not shape:
-        shape = (2, 8, 16, 32, 32)
-
     guidewindow["pedestal_frames"] = u.Quantity(np.zeros(shape, dtype=np.uint16), u.DN, dtype=np.uint16)
     guidewindow["signal_frames"] = u.Quantity(np.zeros(shape, dtype=np.uint16), u.DN, dtype=np.uint16)
     guidewindow["amp33"] = u.Quantity(np.zeros(shape, dtype=np.uint16), u.DN, dtype=np.uint16)
@@ -1295,7 +1250,7 @@ def mk_guidewindow(shape=None, filepath=None):
         return guidewindow
 
 
-def mk_saturation(shape=None, filepath=None):
+def mk_saturation(shape=(4096, 4096), filepath=None):
     """
     Create a dummy Saturation instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -1318,9 +1273,6 @@ def mk_saturation(shape=None, filepath=None):
     meta["reftype"] = "SATURATION"
     saturationref["meta"] = meta
 
-    if not shape:
-        shape = (4096, 4096)
-
     saturationref["dq"] = np.zeros(shape, dtype=np.uint32)
     saturationref["data"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.DN, dtype=np.float32)
 
@@ -1332,7 +1284,7 @@ def mk_saturation(shape=None, filepath=None):
         return saturationref
 
 
-def mk_superbias(shape=None, filepath=None):
+def mk_superbias(shape=(4096, 4096), filepath=None):
     """
     Create a dummy Superbias instance (or file) with arrays and valid values for attributes
     required by the schema.
@@ -1354,9 +1306,6 @@ def mk_superbias(shape=None, filepath=None):
     superbiasref = stnode.SuperbiasRef()
     meta["reftype"] = "BIAS"
     superbiasref["meta"] = meta
-
-    if not shape:
-        shape = (4096, 4096)
 
     superbiasref["data"] = np.zeros(shape, dtype=np.float32)
     superbiasref["dq"] = np.zeros(shape, dtype=np.uint32)