+++ CHANGES.rst
@@ -16,6 +16,9 @@
 
 - Add Level 3 MosaicModel and Resample stnodes, maker utils, factories, and tests. [#163]
 
+- Renamed n_ints to n_groups. Did some shape variable cleanup. [#165]
+
+- Bugfix for the ``amp33`` shape in ``mk_ramp``. [#166]
 
 
 0.14.2 (2023-03-31)